pyproject.toml
[project]
name = "flood-risk"
version = "0.0.1"

# Optional descriptive information for the project.
description = "Determine flooding risk for addresses"
keywords = ["flooding", "risk ", "climate change", "netherlands"]
readme = "README.md"

# Fill in authors and maintainers here.
authors = [
    {name = "Lukas Koning", email = "lukas.koning@afm.nl"},
]
maintainers = [
    {name = "Lukas Koning", email = "lukas.koning@afm.nl"},
]

# Optional classifications.
# See: https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

# Python version and dependencies
requires-python = ">=3.10"
dependencies = [
    "geopy ~= 2.4.1",
    "pandas ~= 2.2.1",
    "rioxarray ~= 0.15.1",
]

[project.optional-dependencies]
# Regular development dependencies.
dev = [
    "isort >= 5.12",
    "black >= 23.10",
    "pylint >= 3.0",
]

[project.urls]
# URL to the DevOps repository.
repository = "https://github.com/LFKoning/flood-risk"

[build-system]
# Using setuptools as build backend.
requires = ["setuptools >= 61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# Place package source code in src folder.
where = ["src"]

[project.scripts]
floodrisk = "flood_risk.cli:main"

# Pylint settings.
# To generate an example: `pylint --generate-toml-config > pylint_example.toml`

[tool.pylint.main]
# Set number of concurrent jobs, use 0 to autodetect.
jobs=0

[tool.pylint.basic]
# Good variable names which should always be accepted.
good-names = ["X", "y", "df", "e", "n"]

[tool.pylint.format]
# Maximum characters on a single line (same as black).
max-line-length = 88

[tool.pylint.messages_control]
# List of messages or message categories to ignore.
disable = ["logging-fstring-interpolation"]